Folders: Remove store
